Working with formsets
Add formsets to the idea form using django-crispy-forms.
The form layout for the translations forms has the main fieldset,
id and DELETE fields that are necessary for recognition of each model
instance and the possibility to remove them from the list. The DELETE
field is actually a checkbox that if checked, deletes the appropriate
item from the database. Also, the form helper for the translation has
form_tag=False, which doesn't generate the <form> tag, and
disable_csrf=True, which doesn't include the CSRF token, because it
have already been defined these in the parent form.

The inlines.js JavaScript does a few things:
For each existing formset form, it initializes its JavaScript-powered
widgets (you could use tooltips, day or color pickers, maps, and so on)
and creates a delete button, which is shown instead of the DELETE
checkbox.
When a delete button is clicked, it checks the DELETE checkbox and
hides the formset form from the user.
When the add button is clicked, it clones the empty form and replaces
"__prefix__" with the next available index, adds the new form
to the list, and initiates JavaScript-powered widgets.