Managing paginated lists
When QuerySet is filtered out, we will create a paginator object
passing QuerySet and the maximal amount of items that we want to show
per page, which is 24 here. Then, we will read the current page number
from the query parameter, page. The next step is to retrieve the current
page object from the paginator. If the page number is not an integer,
we get the first page. If the number exceeds the number of possible
pages, the last page is retrieved. The page object has methods and
attributes necessary for the pagination widget shown in the preceding
screenshot. Also, the page object acts like QuerySet so that we can
iterate through it and get the items from the fraction of the page