Implementing a multilingual search
MultilingualWhooshEngine specifies two custom properties:
    - backend points to MultilingualWhooshSearchBackend, which ensures
    that the items will be indexed for each language given
    in the LANGUAGES setting and put under the associated Haystack
    index location defined in HAYSTACK_CONNECTIONS;
    - query references MultilingualWhooshSearchQuery, whose
    responsibility is to ensure that, when searching for keywords,
    the Haystack connection specific to the current language
    will be used.

Each index has a text field, where full text from a specific language
of a model will be stored. The model for the index is determined
by the get_model() method, the index_queryset() method defines what
QuerySet to index, and the content to search within is defined
as a newline-separated string in the prepare_text() method.

For the template, we have incorporated a few elements of Bootstrap 4
using the out-of-the-box rendering capabilities for forms.