Update base template, create header and footer
The base template contains the <head> and <body> sections of the HTML
document, with all the details that are reused on each page of the
website. Depending on the web design requirements, you can have
additional base templates for different layouts. For example,
we can add the base_simple.html file, which has the same HTML <head>
section and a very minimalistic <body> section, and this can be used
for the login screen, password reset, or other simple pages. You can
have separate base templates for other layouts as well, such
as single-column, two-column, and three-column layouts, where each
extends base.html and overwrites the blocks as needed.

The base template that we created is, by no means, a static unchangeable
template. You can modify the markup structure, or add the elements
you need to it—for example, a template block for body attributes,
a snippet for Google Analytics code, common JavaScript files, the Apple
touch icon for iPhone bookmarks, Open Graph meta tags, Twitter Card
tags, schema.org attributes, and so on. You may also want to define
other blocks, depending on the requirements of your project, and
maybe even wrap the whole content of the body so that you can overwrite
it in a child template.