Using Django Sekizai
Django Sekizai works with the templates included by the {% include %}
template tag, custom template tags that are rendered with templates,
or templates for form widgets. The {% addtoblock %} template tags define
 the Sekizai block that we want to add HTML content to.

When you add something to a Sekizai block, django-sekizai takes care
of including it there only once. This means that you can have multiple
included widgets of the same type, but their CSS and JavaScript will
only be loaded and executed once.